requirements.txt
Django==5.0
djangorestframework==3.14.0
drf-yasg[validation]==1.21.7
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
ipython==8.18.1
requests==2.31.0
pyTelegramBotAPI==4.14.1